# backend/docs/swagger.json
{
    "components": {
        "schemas": {
            "auth.Role": {
                "enum": [
                    "ADMIN",
                    "COLLABORATOR"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "RoleAdmin",
                    "RoleCollaborator"
                ]
            },
            "auth.User": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/auth.Role"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "auth.acceptInvitationRequest": {
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "auth.createInvitationRequest": {
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "auth.loginRequest": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "SessionCookie": {
                "description": "HTTP-only session cookie used to authenticate requests.",
                "in": "cookie",
                "name": "paper_server_session",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Returns the current API server health status.",
        "title": "EnderTex API",
        "version": "1.0"
    },
...
    "paths": {
        "/admin/invitations": {
            "post": {
                "description": "Creates a collaborator invitation and returns the invitation token.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.createInvitationRequest",
                                "summary": "request",
                                "description": "Invitation details"
                            }
                        }
                    },
                    "description": "Invitation details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Created invitation"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Invalid request or invitation"
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Method not allowed"
                    }
                },
                "summary": "Create invitation",
                "tags": [
                    "Administration"
                ]
            }
        },
        "/auth/accept-invitation": {
            "post": {
                "description": "Accepts a valid invitation token and creates the associated user account.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.acceptInvitationRequest",
                                "summary": "request",
                                "description": "Invitation acceptance details"
                            }
                        }
                    },
                    "description": "Invitation acceptance details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.User"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Invalid invitation or request"
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Method not allowed"
                    }
                },
                "summary": "Accept invitation",
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticates a user using email and password.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.loginRequest",
                                "summary": "request",
                                "description": "Login credentials"
                            }
                        }
                    },
                    "description": "Login credentials",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "Login",
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Invalidates the current session and clears the session cookie.",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Method not allowed"
                    }
                },
                "summary": "Logout",
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/auth/me": {
            "get": {
                "description": "Returns the user associated with the current session.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Method not allowed"
                    }
                },
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "summary": "Get current user",
                "tags": [
                    "Authentication"
                ]
            }
        },
...
    },
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "/api"
        }
    ]
}